# public/openapi.yaml
openapi: 3.0.3
info:
  title: VoiceFlow Pro API
  description: |
    VoiceFlow Pro is an AI-powered voice agent platform that helps businesses automate
    customer interactions through intelligent phone calls.

    ## Authentication
    All API endpoints require authentication using Bearer tokens. Include your API key
    in the Authorization header:
    ```
    Authorization: Bearer YOUR_API_KEY
    ```

    ## Rate Limits
    - Read endpoints: 100 requests/minute
    - Write endpoints: 30 requests/minute
    - Voice endpoints: 10 requests/minute

    ## Webhooks
    When a flow is published, it receives a unique webhook URL for receiving inbound calls.
    Webhook requests include a signature header for verification.

  version: 1.0.0
  contact:
    name: VoiceFlow Pro Support
    email: support@voiceflowpro.com
  license:
    name: Proprietary
servers:
  - url: https://api.voiceflowpro.com/v1
    description: Production server
  - url: http://localhost:3000/api
    description: Local development
...
paths:
...
  /analytics:
    get:
      tags:
        - Analytics
      summary: Get analytics data
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
            enum: [7, 30, 90]
          description: Number of days to analyze
      responses:
        '200':
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsData'
...
components:
  schemas:
...
    AnalyticsData:
      type: object
      properties:
        overview:
          type: object
          properties:
            totalCalls:
              type: integer
            avgDuration:
              type: number
            successRate:
              type: number
            totalCost:
              type: number
        callVolume:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              inbound:
                type: integer
              outbound:
                type: integer
        outcomes:
          type: object
          properties:
            completed:
              type: integer
            failed:
              type: integer
            noAnswer:
              type: integer
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []